Move validate_directory_writable to chia.cmds.cmds_util
