+++ chia/_tests/util/test_misc.py
@@ -8,7 +8,7 @@ import anyio
 import pytest
 
 from chia._tests.util.misc import RecordingWebServer
-from chia.cmds.cmds_util import format_bytes, format_minutes
+from chia.cmds.cmds_util import format_bytes, format_minutes, validate_directory_writable
 from chia.types.blockchain_format.program import Program
 from chia.util.errors import InvalidPathError
 from chia.util.ints import uint64
@@ -20,7 +20,6 @@ from chia.util.misc import (
     split_async_manager,
     split_manager,
     to_batches,
-    validate_directory_writable,
 )
 from chia.util.timing import adjusted_timeout, backoff_times
